requirements.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.7.2
pytz==2020.1
Serializer==0.2.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.1
whitenoise==5.2.0
